Refactor for a single source of truth
